refactor: tokenize_checkerをcheckerに移動
