+++ Makefile
@@ -3,10 +3,10 @@
 #                                                         :::      ::::::::    #
 #    Makefile                                           :+:      :+:    :+:    #
 #                                                     +:+ +:+         +:+      #
-#    By: kjikuhar <kjikuhar@student.42tokyo.jp>     +#+  +:+       +#+         #
+#    By: urassh <urassh@student.42.fr>              +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2025/05/18 21:11:39 by kjikuhar          #+#    #+#              #
-#    Updated: 2025/11/04 22:30:29 by kjikuhar         ###   ########.fr        #
+#    Updated: 2025/11/05 13:47:17 by urassh           ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -58,6 +58,7 @@ LIBFT_A		:=	$(LIBFT_DIR)/libft.a
 SRCS_MAND	:=	src/main.c \
 				src/prompt/prompt.c \
 				src/callback/on_input.c \
+				src/checker/tokenize_checker.c \
 				src/tokenize/tokenize.c \
 				src/tokenize/is_operator.c \
 				src/tokenize/state/in_normal.c \
